// microverse-generator/src/interfaces/audioTypes.ts
export type Pattern = {
    time: {
        bpm: number;
        numerator: number;
        denominator: number;
        sequencerRate: number;
        listenToMidiClock: boolean;
    },
    patternArr: PatternCell[]; 
    patternArrName: string;
    scaleIntervalName?: string;
    microtoneIntervalName?: string; 
    intervalArr?: string[];
    microtoneArr?: string[]; 
}

// INSTRUMENTS ****************************
// ****************************************

export type InstrumentNameVal = {
    name: string; 
    value: any;
    type: string;
    effectsString: string;
    effectsHelper: any;

}

export type STKInstruments = {
    Clarinet: EffectsSettings;
    StifKarp: EffectsSettings;
    Sitar: EffectsSettings;
    FrencHrn: EffectsSettings;
    Moog: EffectsSettings;
    Rhodey: EffectsSettings;
    Saxofony: EffectsSettings;
    Mandolin: EffectsSettings;
    BandedWg: EffectsSettings;
    BlowBtl: EffectsSettings;
    BlowHole: EffectsSettings;
    Bowed: EffectsSettings;
    Brass: EffectsSettings;
    Flute: EffectsSettings;
    ModalBar: EffectsSettings;
    Shakers: EffectsSettings;
    VoicForm: EffectsSettings;
    B3: EffectsSettings;
    HevyMetl:EffectsSettings;
    HonkeyTonk:EffectsSettings;
    FMVoices:EffectsSettings;
    KrstlChr:EffectsSettings;
    PercFlut:EffectsSettings;
    TubeBell:EffectsSettings;
    Wurley:EffectsSettings;
}


// SOURCES *********************************
// *****************************************

export type Source = {
    masterVolume: number;
    masterPan: number;
    detune: number;
    effects: Effects;
    effectsString: string;
    pattern: Pattern[];
    arpeggiateOn: boolean;
    instruments?: STKInstruments;
    files?: File[];
    active: boolean;
    isEditing: boolean;
    isMuted: boolean;
    isSolo: boolean;
    // samplerRate: number;
}


export type Sources = {
    osc1: Source;
    // osc2: Source;
    stk1: Source;
    sampler: Source;
    audioin: Source;
}
// ...
export type Effects = {
    WinFuncEnv: EffectsSettings;
    // WinFunc: EffectsSettings;
    ExpEnv: EffectsSettings;
    WPDiodeLadder: EffectsSettings;
    WPKorg35: EffectsSettings;
    Modulate: EffectsSettings;
    Delay: EffectsSettings;
    DelayA: EffectsSettings;
    DelayL: EffectsSettings;
    ExpDelay: EffectsSettings;
    Elliptic: EffectsSettings;
    Spectacle: EffectsSettings;
    Gain: EffectsSettings;
    Bitcrusher: EffectsSettings;
    FoldbackSaturator: EffectsSettings;
    Echo: EffectsSettings;
    Chorus: EffectsSettings;
    PitShift: EffectsSettings;
    AmbPan3: EffectsSettings;
    JCRev: EffectsSettings;
    NRev: EffectsSettings;
    PRCRev: EffectsSettings;
    GVerb: EffectsSettings;
    // ADSR: EffectsSettings;
    PowerADSR: EffectsSettings;
    KasFilter: EffectsSettings;
    Multicomb: EffectsSettings;
    PitchTrack: EffectsSettings;
    Sigmund: EffectsSettings;
    SndBuf: EffectsSettings;
    LiSa: EffectsSettings;
}